python-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
httpx==0.27.0
tenacity==8.2.3
pydantic==2.6.4
python-dotenv==1.0.1
python-json-logger==2.0.7
pytest==7.4.4
pytest-asyncio==0.23.5
respx==0.20.2
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
opentelemetry-instrumentation-logging==0.45b0
pymongo==4.6.3
mongomock==4.1.2
networkx==3.2.1

# ML/AI Dependencies
torch==2.2.0
torch-geometric==2.5.0
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
mlflow==2.10.2
transformers==4.38.1
sentence-transformers==2.5.1

# Phase 2: Multi-Modal Dependencies
rdkit==2023.9.4  # Chemical structure analysis

# Phase 3: Advanced Knowledge Graph Dependencies
node2vec==0.4.6  # Node2Vec graph embeddings
gensim==4.3.2  # Word2Vec for node2vec
umap-learn==0.5.5  # UMAP dimensionality reduction (optional)

# Phase 4: Enhanced ML Models Dependencies
scipy==1.12.0  # Statistical functions for uncertainty quantification

# Neo4j Graph Database
neo4j==6.0.2  # Neo4j Python driver for graph database operations
